+++ lib/widgets/map/geomap_settings.dart
@@ -16,8 +16,8 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
 
 import 'package:geopod/models/place.dart';
-import 'package:geopod/services/map_settings_pod.dart';
 import 'package:geopod/services/map_settings_service.dart';
+import 'package:geopod/services/map_viewport_store.dart';
 import 'package:geopod/services/places/encrypted_places_service.dart';
 import 'package:geopod/services/places_service.dart' show PlacesCacheManager;